+++ js/sidebar.js
@@ -114,11 +114,23 @@ function createSearch() {
   }
 }
 
+// Cuts the name of a file / folder off after a certain number of characters.
+let sliceIndex;
+function sliceName(name) {
+  return name.slice(0, sliceIndex) + "...";
+}
+
 // Creating Folder to be rendered on the Sidebar
 function createFolder(entry) {
   const folder = document.createElement("div");
   folder.classList.add("folder");
-  folder.textContent = entry.name;
+
+  let folderName = entry.name;
+  if (folderName.length > sliceIndex) {
+    folderName = sliceName(folderName);
+  }
+
+  folder.textContent = folderName;
   return folder;
 }
 
@@ -126,9 +138,24 @@ function createFolder(entry) {
 function createFile(entry, previousEntry) {
   const file = document.createElement("div");
   file.classList.add("file");
-  file.textContent = entry.name.slice(0, -5);
+
+  let fileName = entry.name.slice(0, -5);
+  if (fileName.length > sliceIndex) {
+    fileName = sliceName(fileName);
+  }
+
+  file.textContent = fileName;
 
   file.addEventListener("click", async (e) => {
+    const urlParams = new URLSearchParams(window.location.search);
+    const pathParam = urlParams.get("path");
+    const documentParam = urlParams.get("document");
+
+    // Cancels document loading if the current document is already opened.
+    if (pathParam === previousEntry && documentParam === entry.name) {
+      return;
+    }
+
     const response = await fetch(
       `api/documents/getFile?folderPath=${previousEntry}&name=${entry.name}`,
       {
@@ -142,7 +169,7 @@ function createFile(entry, previousEntry) {
         `?path=${previousEntry}&document=${entry.name}`,
       ); // Sets the Query Parameters into the URL
       const fileData = await response.json(); // Data from GET request
-      loadAutosave(fileData, entry.name, previousEntry); // Checks for an Autosave, which then loads the document.
+      loadAutosave(fileData, entry.name, previousEntry); // Checks for an autosave, which then loads the document.
       setState("currentDocument", previousEntry + entry.name);
     } else if (response.status === 404) {
       createErrorModal("Couldn't find the File you were looking for.");
@@ -410,6 +437,8 @@ function renderEntries(entries, indentlevel, previousEntry) {
 }
 
 export async function buildSidebar() {
+  // Setting variables
+  sliceIndex = getState("sliceIndex");
   // Getting data
   const response = await fetch("api/documents", {
     method: "GET",